requirements/requirements.txt
asgiref==3.2.7
attrs==19.3.0
bcrypt==3.1.7
cached-property==1.5.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
Django==3.0.5
django-ckeditor==5.9.0
django-js-asset==1.2.2
docker==4.2.0
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
idna==2.9
jsonschema==3.2.0
paramiko==2.7.1
psycopg2==2.8.5
pycparser==2.20
PyNaCl==1.3.0
pyrsistent==0.16.0
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
texttable==1.6.2
urllib3==1.25.9
websocket-client==0.57.0
